Fix incorrect current-buffer bug in 'posframe-show'
The function 'font-at' (called from 'posframe--get-font-height')
signals an error when the current buffer is different from the one
displayed in the selected window.

Make sure the correct buffer is current.

(This is not theoretical, I have actually bumped into the error
myself. It is not guaranteed that selected-window matches
current-buffer during 'posframe-show' invocation.)