Make the buffer in posframe dedicated
- I am trying to use posframe to create a floating control for dap-mode which
will have Step in/Step out/Continue, etc buttons (see screenshot). The problem
that I am facing is that when posframe-show which is called when the new set of
buttons is enabled is performed when current-frame the posframe itself. In order
to solve that in our code I used

``` elisp
(when (eq (selected-frame) pos-frame)
  (select-frame (frame-parent pos-frame)))
```

But when I do that from time to time the buffer in the posframe chamges. The
following piece of code fixes the issue.